Support High DPI screens
This is a very basic implementation of High DPI screens for Julius. It
simply passes the SDL_WINDOW_ALLOW_HIGHDPI flag to SDL_CreateWindow().

This means that the OS will still scale window resolutions: so, for
example, on a 3200×1800px screen which has an OS scaling factor of 2×,
Julius will run fullscreen at what appears to be 1600×900, and the
windowed modes will all be scaled 2×. Julius will still show the scale
as 100%. Setting the in-game scale to 50% on such a machine will give an
actually 3200×1800px display, and it will look much crisper than it
would without this change.

The only visible difference will be that the scaling will be smoother,
as Julius will be doing all of the scaling work (through
SDL_RenderSetLogicalSize()), rather than scaling to a half-resolution
buffer, then letting the OS scale it again.