Implement dynamic support for new resources (#768)
Most changes were made to savegames to support the future addition of new resources without having to mess with the save file format too much.

The bulk of the changes are in game/resource.c and game/resource.h.

The savegame format now has an extra "resource version" field, which is independent from the actual save version.

Upon load, saves are checked for the resource version. Depending on the version, the game "remaps" the resource ID's from the save file to the new expected ID's.
If a save file doesn't have a resource version, it will map the ID's as if they were from C3.

Resource information is also properly listed on game/resource.c, so the game can get the proper image/text/information data from a resource.

The old "inventory" system has been removed and now resources are directly used. To distinguish between regular and inventory resources, the resource information has a flag indicating whether that resource can be inventoried. This allows dynamically adding resources to the inventory.

Although the engine now supports more than 16 resources, the UI doesn't yet. Scrollbars still need to be added and a proper way to display the house, granary and warehouse stores in the building window needs to be implemented.

This needs a lot of testing.