Change aqueduct routing to use more robust map_can_place_aqueduct_on_highway() check. Just needed to modify it to allow it to count routed tiles as aqueducts. (#828)
For posterity, the issue with the old function was that the first tile wasn't checked to see if it should be skipped. This was kind of a hack put in to allow for clicking and dragging across the highway if you started on the highway. Otherwise the aqueduct would think it needs to be skipped even though it can go across in a straight line using the first tile.

Since the first checked tile was close enough to the end of the highway to clear it, it would get a score (as well as the next tile) and the tile after the highway ended would get enqueued. This could be fixed by moving the "clear" check up into the "skip" check, but then it doesn't account for corners. Since map_can_place_aqueduct_on_highway() already handles all of these cases, I've changed the routing code to use it instead.